+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ### v4.0.0 (in development)
 
-- Added `homeDirectory`, `cachesDirectory` and `tmpDirectory` to `FileManager` extension.
+- Added `homeDirectory`, `cachesDirectory`, `appSupportDirectory` and `tmpDirectory` to `FileManager` extension.
 - Added `base64Encoded` and `base64Decoded` as a string extension.
 - Added `RgbaColor` inclusive initializer and return value for `UIColor`.
 - Added `UIColor.asData()` and `Data.asUIColor()` extensions to archive and unarchive UIColors.